added object to hold state and response

+++ src/Rafty/Log/LogEntry.cs
@@ -4,7 +4,7 @@
 
     public class LogEntry
     {
-        public LogEntry(string commandData, Type type, int term, int currentCommitIndex)
+        public LogEntry(string commandData, Type type, long term, long currentCommitIndex)
         {
             CommandData = commandData;
             Type = type;
@@ -14,7 +14,7 @@
 
         public string CommandData { get; private set; }
         public Type Type { get; private set; }
-        public int Term { get; private set; }
-        public int CurrentCommitIndex { get; private set; }
+        public long Term { get; private set; }
+        public long CurrentCommitIndex { get; private set; }
     }
 }
